Add Authentication to Handlers

+++ CompanyTestProject/CompanyTestProject.Application/Features/Product/Commands/Delete/DeleteProductCommandHandler.cs
@@ -1,21 +1,17 @@
-﻿using AutoMapper;
-using CompanyTestProject.Application.Repositories;
+﻿using CompanyTestProject.Application.Repositories;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace CompanyTestProject.Application.Features.Product.Commands.Delete
 {
     public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand, Unit>
     {
         private readonly IProductRepository _ProductRepository;
+        private readonly IUserProductRepository _UserProductRepository;
 
-        public DeleteProductCommandHandler(IProductRepository productRepository)
+        public DeleteProductCommandHandler(IProductRepository productRepository, IUserProductRepository userProductRepository)
         {
             _ProductRepository = productRepository;
+            _UserProductRepository = userProductRepository;
         }
         public async Task<Unit> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
@@ -23,6 +19,11 @@ namespace CompanyTestProject.Application.Features.Product.Commands.Delete
             if (product == null)
                 throw new Exception("This product doesn't exist");
 
+            var userProducts = await _ProductRepository.GetByUserId(request.UserId);
+            if (!userProducts.Exists(x => x.Id == product.Id))
+                throw new Exception("You don't have permission to delete this product");
+
+            await _UserProductRepository.Delete(product.Id);
             await _ProductRepository.Delete(product);
 
             return Unit.Value;
